# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: DigiHostel API
  version: 0.1.0
  description: >
    Source contract for the DigiHostel API (ADR-007). Health check plus the
    first two real business workflows, Parent Leave Approval and Student Leave
    Request creation/viewing (docs/leave-approval-workflow.md).
servers:
  - url: /api/v1
paths:
...
  /emergencies:
    get:
      operationId: listEmergencies
      tags: [emergencies]
      summary: >
        Staff-only: server-side paginated/filtered incident queue (Phase 4,
        Prompt 10 — Emergency Operations Center)
      description: >
        Requires an AAL2 staff session with role reception_warden,
        hostel_admin, or super_admin — library_incharge has no grant.
        reception_warden/hostel_admin are hostel-scoped to their own
        assigned hostel, resolved entirely server-side; super_admin is
        unscoped. Only the 8 EOC emergency categories are ever returned —
        never the separate, still-unbuilt checkpoint-monitoring domain's
        `missed_checkpoint`/`manual_flag` rows. `q` matches a
        case-insensitive PREFIX against the linked student's full_name OR
        roll_number, mirroring `searchStudents`.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string, maxLength: 200 }
        - name: category
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/EmergencyCategory" }
        - name: severity
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/EmergencySeverity" }
        - name: status
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/EmergencyStatus" }
        - name: activeOnly
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: sortBy
          in: query
          required: false
          schema: { type: string, enum: [reportedAt, severity], default: reportedAt }
        - name: sortDir
          in: query
          required: false
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: A page of incidents within the caller's authorized scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyList"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: reportEmergency
      tags: [emergencies]
      summary: >
        Staff-only: log a new incident report (Phase 4, Prompt 10)
      description: >
        A genuine, honest reception-desk capability — a staff member records
        an incident they became aware of (phone call, walk-in report, direct
        observation). The intended Student Application "Emergency Trigger"
        producer described in the product roadmap does not exist anywhere in
        this repository (apps/student-mobile is an unmodified template
        scaffold) — this endpoint is not a substitute for that pipeline, it
        is the real staff-attestation path this table's original design
        already established (mirrors the existing `manual_flag` concept).
        `rollNumber` resolves the student server-side, hostel-scoped
        identically to every other staff route — never a client-supplied
        studentId/hostelId/staffId.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rollNumber, category, severity, description]
              additionalProperties: false
              properties:
                rollNumber: { type: string, minLength: 1, maxLength: 100 }
                category: { $ref: "#/components/schemas/EmergencyCategory" }
                severity: { $ref: "#/components/schemas/EmergencySeverity" }
                description: { type: string, minLength: 1, maxLength: 4000 }
      responses:
        "201":
          description: Incident reported — status `open`, unassigned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Roll number does not resolve to a student in the caller's scope.
  /emergencies/statistics:
    get:
      operationId: getEmergencyStatistics
      tags: [emergencies]
      summary: >
        Staff-only: server-derived active-incident counts for the EOC's
        statistics strip (Phase 4, Prompt 10)
      description: >
        Every count is a fresh aggregate query over the caller's own
        authorized scope — never a client-supplied or cached number.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyStatistics"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /emergencies/{incidentId}:
    get:
      operationId: getEmergency
      tags: [emergencies]
      summary: >
        Staff-only: single incident detail with operational timeline (Phase
        4, Prompt 10)
      description: >
        Returns 404 for both "no such incident" and "exists but outside the
        caller's hostel scope" — anti-enumeration, matching every other
        staff detail route in this API.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: The incident and its full timeline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyDetail"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /emergencies/{incidentId}/notes:
    post:
      operationId: addEmergencyNote
      tags: [emergencies]
      summary: "Staff-only: append an operational note to the incident timeline"
      description: >
        Allowed for any non-`closed` status. Immutable once written — this
        table (security_incident_events) mirrors leave_approval_events's
        append-only discipline exactly.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              additionalProperties: false
              properties:
                note: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        "201":
          description: Note recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth-issued JWT (ADR-014).
  parameters:
...
    IncidentId:
      name: incidentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    LeaveRequestStatus:
      type: string
      description: >
        Full state vocabulary already defined by the schema
        (packages/db/src/schema/enums.ts, leave_request_status) — this
        endpoint set only ever produces approved/rejected transitions today;
        the escalation-notification states are included because they are
        part of the same enum, not because this task creates them.
      enum:
        - pending
        - father_notified
        - mother_notified
        - guardian_notified
        - approved
        - rejected
        - in_app_call
        - manual_verification
        - expired
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            currentStatus:
              $ref: "#/components/schemas/LeaveRequestStatus"
            requestId:
              type: string
              description: >-
                Server-generated per-request identifier (F-07 production
                observability). Present on unexpected (5xx) and
                framework-rejected (4xx) errors; hand this value to support
                to correlate a client-visible failure with server-side logs.
                Not present on typed domain errors that routes map to a
                response themselves (e.g. leave-request 404/409).
...
    EmergencyCategory:
      type: string
      description: >
        The 8 categories added to security_incident_type for the Emergency
        Operations Center (Phase 4, Prompt 10). The table's original two
        values (missed_checkpoint, manual_flag) belong to the separate,
        still-unbuilt Digital Library Pass checkpoint-monitoring domain and
        are never returned/accepted here.
      enum:
        - medical
        - personal_safety
        - fire
        - security_threat
        - violence
        - infrastructure
        - harassment
        - other
    EmergencySeverity:
      type: string
      enum: [critical, high, medium, low, informational]
    EmergencyStatus:
      type: string
      description: >
        The EOC's own 5-state lifecycle. `escalated` (a value that exists on
        the underlying column for the separate checkpoint-monitoring domain)
        is deliberately not part of this API's vocabulary — no concrete EOC
        requirement justifies exposing it.
      enum: [open, acknowledged, in_progress, resolved, closed]
    EmergencyEventType:
      type: string
      enum: [created, acknowledged, response_started, note_added, resolved, closed]
    EmergencyEvent:
      type: object
      description: >
        One immutable security_incident_events row — the EOC's own
        operational timeline (mirrors LeaveApprovalEvent's shape). Unlike
        LeaveApprovalEvent, actor identity (a staff member's name) IS
        included: this domain is staff-only end to end, so showing which
        colleague acted carries none of the student/parent-facing
        disclosure concern that discipline exists to prevent elsewhere.
      required: [id, eventType, note, actorStaffName, occurredAt]
      properties:
        id: { type: string, format: uuid }
        eventType: { $ref: "#/components/schemas/EmergencyEventType" }
        note: { type: string, nullable: true }
        actorStaffName: { type: string, nullable: true }
        occurredAt: { type: string, format: date-time }
    EmergencyListItem:
      type: object
      description: >
        Deliberately minimal student identification (name/roll number/
        hostel/room) — reused DISPLAY fields only, never a duplicated
        guardian/leave query. Use "Open Student Profile" (the existing
        Student Operations Center) for anything more.
      required:
        - id
        - studentId
        - studentFullName
        - studentRollNumber
        - hostelId
        - hostelName
        - roomNumber
        - category
        - severity
        - status
        - reportedAt
        - assignedStaffId
        - assignedStaffName
      properties:
        id: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        studentFullName: { type: string }
        studentRollNumber: { type: string }
        hostelId: { type: string, format: uuid, nullable: true }
        hostelName: { type: string, nullable: true }
        roomNumber: { type: string, nullable: true }
        category: { $ref: "#/components/schemas/EmergencyCategory" }
        severity: { $ref: "#/components/schemas/EmergencySeverity" }
        status: { $ref: "#/components/schemas/EmergencyStatus" }
        reportedAt: { type: string, format: date-time }
        assignedStaffId: { type: string, format: uuid, nullable: true }
        assignedStaffName: { type: string, nullable: true }
    EmergencyList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/EmergencyListItem" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    EmergencyDetail:
      allOf:
        - $ref: "#/components/schemas/EmergencyListItem"
        - type: object
          required: [description, resolvedAt, closedAt, timeline]
          properties:
            description: { type: string, nullable: true }
            resolvedAt: { type: string, format: date-time, nullable: true }
            closedAt: { type: string, format: date-time, nullable: true }
            timeline:
              type: array
              items: { $ref: "#/components/schemas/EmergencyEvent" }
    EmergencyStatistics:
      type: object
      description: >
        Server-derived, active-incident counts — every field a fresh
        aggregate query over the caller's own authorized scope, never a
        client-supplied or cached number.
      required: [active, critical, open, acknowledged, inProgress, resolvedToday]
      properties:
        active: { type: integer }
        critical: { type: integer }
        open: { type: integer }
        acknowledged: { type: integer }
        inProgress: { type: integer }
        resolvedToday: { type: integer }
...
  responses:
    Unauthenticated:
      description: Missing, malformed, or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Forbidden:
      description: Authenticated but not authorized for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: >
        Leave request does not exist, or exists but the caller has no
        relationship to its student (deliberately indistinguishable —
        anti-enumeration).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Conflict:
      description: Leave request is not in a state that admits this decision
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    ValidationError:
      description: Request body/params failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"